feat: componentize team stats for use on referrals page
